chore: rename collector
